drishti-mother-child/src/test/java/org/ei/drishti/service/reporting/rules/IsChildFemaleRuleTest.java
package org.ei.drishti.service.reporting.rules;

import org.ei.drishti.util.SafeMap;
import org.junit.Before;
import org.junit.Test;

import java.util.Map;

import static org.ei.drishti.util.EasyMap.create;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;


public class IsChildFemaleRuleTest {

    private IsChildFemaleRule rule;

    @Before
    public void setUp() {
        rule = new IsChildFemaleRule();
    }

    @Test
    public void shouldReturnFalseWhenChildIsNotFemale() {
        Map<String, String> reportData = create("gender", "male")
                .map();
        boolean didRuleSucceed = rule.apply(new SafeMap(reportData));

        assertFalse(didRuleSucceed);

        reportData = create("gender", "")
                .map();
        didRuleSucceed = rule.apply(new SafeMap(reportData));

        assertFalse(didRuleSucceed);
    }

    @Test
    public void shouldReturnTrueWhenChildIsFemale() {
        Map<String, String> reportData = create("gender", "female")
                .map();
        boolean didRuleSucceed = rule.apply(new SafeMap(reportData));

        assertTrue(didRuleSucceed);
    }
}
